+++ src/stores/todoListStore.js
@@ -1,18 +1,36 @@
-import { writable } from 'svelte/store';
+import { persisted } from 'svelte-local-storage-store';
 
-export const todoList = writable([]);
+export const todoList = persisted('pinBoard', {
+  tasks: [],
+  itemsNeeded: [],
+});
+
+export function addItemNeeded(item) {
+  todoList.update((todo) => {
+    if (todo.itemsNeeded.find((i) => i.id === item.id)) {
+      return todo;
+    }
+    return { ...todo, itemsNeeded: [...todo.itemsNeeded, item] };
+  });
+}
+
+export function removeItemNeeded(id) {
+  todoList.update((todo) => {
+    return { ...todo, itemsNeeded: todo.itemsNeeded.filter((i) => i.id !== id) };
+  });
+}
 
 export function addTask(task) {
   todoList.update((todo) => {
-    if (todo.find((t) => t.id === task.id)) {
+    if (todo.tasks.find((t) => t.id === task.id)) {
       return todo;
     }
-    return [...todo, task];
+    return { ...todo, tasks: [...todo.tasks, task] };
   });
 }
 
 export function removeTask(id) {
   todoList.update((todo) => {
-    return todo.filter((task) => task.id !== id);
+    return { ...todo, tasks: todo.tasks.filter((t) => t.id !== id) };
   });
 }